+++ out/test/extension.test.js
@@ -591,8 +591,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             outputFormat: "png - Portable Network Graphics"
         };
         let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
-        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=$(wslpath \'c:\\pov\\teapot\\output\\teapot.png\') Library_Path=$(wslpath 'c:\\Users\\myuser\\Documents\\POVRay\\include\\') Dither=on");
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
+        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=$(wslpath \'" + settings.outputPath + "')/ Library_Path=$(wslpath 'c:\\Users\\myuser\\Documents\\POVRay\\include\\') Dither=on");
     });
     test("buildRenderOptions_win32_bash_docker", function () {
         let context = {
@@ -618,8 +618,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_win32_bash_docker_filename_space", function () {
@@ -646,8 +645,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\tea pot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " \"'\"tea\\ pot.pov\"'\" -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_win32_powershell", function () {
@@ -674,9 +672,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
-        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=c:\\pov\\teapot\\output\\teapot.png\ Library_Path=c:\\Users\\myuser\\Documents\\POVRay\\include\\ Dither=on | Out-Null");
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
+        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=" + settings.outputPath + " Library_Path=c:\\Users\\myuser\\Documents\\POVRay\\include\\ Dither=on | Out-Null");
     });
     test("buildRenderOptions_win32_powershell_docker", function () {
         let context = {
@@ -702,8 +699,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_win32_powershell_docker_filename_space", function () {
@@ -730,8 +726,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\tea pot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " '''${fileBasename}''' -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_win32_cmd", function () {
@@ -758,9 +753,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
-        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=c:\\pov\\teapot\\output\\teapot.png\ Library_Path=c:\\Users\\myuser\\Documents\\POVRay\\include\\ Dither=on");
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
+        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=" + settings.outputPath + " Library_Path=c:\\Users\\myuser\\Documents\\POVRay\\include\\ Dither=on");
     });
     test("buildRenderOptions_win32_cmd_docker", function () {
         let context = {
@@ -786,8 +780,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_win32_cmd_docker_filename_space", function () {
@@ -814,8 +807,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "c:\\pov\\teapot\\output\\tea pot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " '\"tea pot.pov\"' -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_linux", function () {
@@ -842,9 +834,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
-        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/pov/teapot/out/teapot.png Library_Path=/Users/myuser/Documents/POVRay/include/ Dither=on");
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
+        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=" + settings.outputPath + " Library_Path=/Users/myuser/Documents/POVRay/include/ Dither=on");
     });
     test("buildRenderOptions_linux_docker", function () {
         let context = {
@@ -870,8 +861,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/output/teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_darwin", function () {
@@ -898,9 +888,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
-        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/pov/teapot/out/teapot.png Library_Path=/Users/myuser/Documents/POVRay/include/ Dither=on");
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
+        assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=" + settings.outputPath + " Library_Path=/Users/myuser/Documents/POVRay/include/ Dither=on");
     });
     test("buildRenderOptions_darwin_docker", function () {
         let context = {
@@ -926,8 +915,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " ${fileBasename} -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     test("buildRenderOptions_darwin_docker_filename_space", function () {
@@ -954,8 +942,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/tea pot.png";
-        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, outFilePath, context);
+        let povrayExe = povrayExtension.buildRenderOptions(settings, fileInfo, context);
         assert.equal(povrayExe, " \"'\"tea\\ pot.pov\"'\" -D Width=1024 Height=768 Output_File_Name=/output/ Dither=on");
     });
     // getDimensionOptions() unit tests
@@ -1050,9 +1037,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.getOutputPathOption(settings, outFilePath, context);
-        assert.equal(povrayExe, " Output_File_Name=$(wslpath \'/pov/teapot/out/teapot.png\')");
+        let povrayExe = povrayExtension.getOutputPathOption(settings, context);
+        assert.equal(povrayExe, " Output_File_Name=$(wslpath \'" + settings.outputPath + "\')/");
     });
     test("getOutputPathOption_docker", function () {
         let context = {
@@ -1078,8 +1064,7 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.getOutputPathOption(settings, outFilePath, context);
+        let povrayExe = povrayExtension.getOutputPathOption(settings, context);
         assert.equal(povrayExe, " Output_File_Name=/output/");
     });
     test("getOutputPathOption_win32_powershell", function () {
@@ -1106,9 +1091,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.getOutputPathOption(settings, outFilePath, context);
-        assert.equal(povrayExe, " Output_File_Name=/pov/teapot/out/teapot.png");
+        let povrayExe = povrayExtension.getOutputPathOption(settings, context);
+        assert.equal(povrayExe, " Output_File_Name=" + settings.outputPath);
     });
     test("getOutputPathOption_linux", function () {
         let context = {
@@ -1134,9 +1118,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.getOutputPathOption(settings, outFilePath, context);
-        assert.equal(povrayExe, " Output_File_Name=/pov/teapot/out/teapot.png");
+        let povrayExe = povrayExtension.getOutputPathOption(settings, context);
+        assert.equal(povrayExe, " Output_File_Name=" + settings.outputPath);
     });
     test("getOutputPathOption_darwin", function () {
         let context = {
@@ -1162,9 +1145,8 @@ suite("VSCode-POVRay Extension Unit Tests", function () {
             useDockerImage: "jmaxwilson/povray",
             outputFormat: "png - Portable Network Graphics"
         };
-        let outFilePath = "/pov/teapot/out/teapot.png";
-        let povrayExe = povrayExtension.getOutputPathOption(settings, outFilePath, context);
-        assert.equal(povrayExe, " Output_File_Name=/pov/teapot/out/teapot.png");
+        let povrayExe = povrayExtension.getOutputPathOption(settings, context);
+        assert.equal(povrayExe, " Output_File_Name=" + settings.outputPath);
     });
     // getLibraryPathOption() unit tests
     test("getLibraryPathOption_win32_bash", function () {